requirements.txt
asgiref==3.4.1
Django==3.2.7
django-admin==2.0.1
django-background-tasks==1.2.5
django-compat==1.0.15
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-six==1.0.4
mysqlclient==2.0.3
pdf2image==1.16.0
Pillow==8.3.2
pytesseract==0.3.8
pytz==2021.1
screen==1.0.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.2
xlwt==1.3.0
